requirements.txt
django
dj-static
dj-database-url
gunicorn
psycopg2
python-decouple